🚚 모달 상수 constants 폴더에 파일 생성 및 이동
